Add: authentication and related changes in frontend

+++ frontend/src/component/WorkoutDetails.js
@@ -1,13 +1,22 @@
 import React from "react";
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ item }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch(`/api/workouts/${item?._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user?.token}`,
+      },
     });
 
     const res = await response.json();
